pyproject.toml
[tool.poetry]
name = "fastapi-async"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "fastapi_async"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.95.1"
sqlalchemy = {extras = ["asyncio"], version ="^2.0.12"}
asyncpg = "^0.27.0"
python-jose = "^3.3.0"
passlib = "^1.7.4"
python-multipart = "^0.0.6"
uvicorn = "^0.22.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
